Estrutura organizada e dependências configuradas
