+++ .gitignore
@@ -1,0 +1,6 @@
+node_modules/
+dist/
+.env
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
