+++ eslint.config.cjs
@@ -1,0 +1,34 @@
+const js = require("@eslint/js");
+const reactPlugin = require("eslint-plugin-react");
+const reactHooks = require("eslint-plugin-react-hooks");
+const globals = require("globals");
+
+module.exports = [
+  js.configs.recommended,
+  {
+    files: ["src/**/*.{js,jsx}"],
+    languageOptions: {
+      ecmaVersion: "latest",
+      sourceType: "module",
+      parserOptions: {
+        ecmaFeatures: { jsx: true },
+      },
+      globals: {
+        ...globals.browser,
+      },
+    },
+    settings: {
+      react: {
+        version: "detect",
+      },
+    },
+    plugins: {
+      react: reactPlugin,
+      "react-hooks": reactHooks,
+    },
+    rules: {
+      ...reactPlugin.configs.recommended.rules,
+      "react/no-unescaped-entities": "off",
+    },
+  },
+];
